Restructure of the fixture + new files for Ingredients blending
