+++ data-qa/delphi/fixtures/calculators/PHcalculatorTestData.json
@@ -1,5 +1,5 @@
 {
-  "uri": "/ph_calculator",
+  "uri": "/interpret/ph_calculator",
   "variants": {
     "liquid": {
       "request": {
